Fix 64 bit warnings.
* include/fct.h (fctkern__cl_parse): Use size_t instead of int to remove
some nuisance warnings on a 64 bit compiler.

+++ include/fct.h
@@ -1995,8 +1995,8 @@ static int
 fctkern__cl_parse(fctkern_t *nk)
 {
     int status =0;
-    int num_params =0;
-    int param_i =0;
+    size_t num_params =0;
+    size_t param_i =0;
     if ( nk == NULL )
     {
         return 0;
